refactor: move Lean from tools to runtimes
Lean is a language runtime, not a tool (like edit, bash, memory).
Refactor to align with gh-aw's `runtimes:` front matter model:

- New `src/runtimes/` module alongside `src/tools/` and `src/safeoutputs/`
- `src/runtimes/lean.rs`: LeanRuntimeConfig, LeanOptions, install/prompt generators
- `src/runtimes/mod.rs`: module documentation
- `RuntimesConfig` struct in types.rs with `runtimes` field on FrontMatter
- Front matter changes from `tools: lean:` to `runtimes: lean:`
- All compiler references updated (standalone.rs, common.rs)
- AGENTS.md: new "Runtimes Configuration" section, separate from Tools
- Example updated
